requirements.txt
openai>=1.30.0
httpx>=0.27.0
pydantic>=2.8.0
PyYAML>=6.0.1
APScheduler>=3.10.4
Jinja2>=3.1.4
Pillow>=10.4.0
moviepy>=1.0.3
